🐛 FIX: clearing of stream when API have an error
