+++ client/reader/stream/error.tsx
@@ -1,9 +1,6 @@
 import { recordTracksEvent } from '@automattic/calypso-analytics';
 import { useTranslate } from 'i18n-calypso';
-import { useEffect } from 'react';
 import EmptyContent from 'calypso/components/empty-content';
-import { useDispatch } from 'calypso/state';
-import { errorNotice } from 'calypso/state/notices/actions';
 
 /**
  * Props for the StreamError component.
@@ -13,31 +10,10 @@ import { errorNotice } from 'calypso/state/notices/actions';
 interface StreamErrorProps {
 	onTryAgain?: () => void;
 	streamKey: string;
-	error: {
-		message: string;
-	};
 }
 
-export const StreamError = ( { onTryAgain, streamKey, error }: StreamErrorProps ) => {
+export const StreamError = ( { onTryAgain, streamKey }: StreamErrorProps ) => {
 	const translate = useTranslate();
-	const dispatch = useDispatch();
-
-	useEffect( () => {
-		if ( ! error.message ) {
-			return;
-		}
-
-		dispatch(
-			errorNotice( translate( 'Stream error: %s', { args: error.message } ), { duration: 3000 } )
-		);
-	}, [] ); // eslint-disable-line react-hooks/exhaustive-deps
-
-	useEffect( () => {
-		recordTracksEvent( 'calypso_reader_stream_error', {
-			stream_key: streamKey,
-			path: window.location.pathname,
-		} );
-	}, [ streamKey ] );
 
 	const handleTryAgain = () => {
 		recordTracksEvent( 'calypso_reader_stream_error_try_again', {
